Initial UI commit
Added index.html for the main screen. The styling is mostly in the same file.
Added uploadFiles.html for the second screen. This page will be used to upload files and pick save locations. It currently allows upload.

Added PowerPoint and testing images in Design folder. These aren't all text as they aren't being tested by the AI, they are jsut used to see how they are displayed with the UI.