Combine AI with UI
Modified uploadFiles.html to pass the document name as well as the
images. The document name is the name of the folder that contains the
images to be converted.

Modified ai.py to return the text that is extracted from the image.

Modified __main__.py to have a folder containing the images to be
converted and a folder that will contain a history of the converted
images and their corresponding output.