Add and link converting.html, bug fixes
converting.html is based off the initial design
 - Previews image and converted text
 - Converted text auto-scrolls to the bottom when new text is added
 - Progress bars reflect the number of files and documents processed
 - There is a progress bar for the current document and overall
 - Progress bar uses CSS variables that are updated via JavaScript
 - Under progress bars, short statuses display

Bug fixes
 - Fixed bug where the old toConvert directory wasn't removed. There was
   a typo.
 - Fixed bug where the file types for picking the save location weren't
   loading

Removed UI testing images and replaced them with screenshots of digital
text for testing. This is because future testing will also test the
conversion process and not just the inputting process.

Changed the name of the stored image file to incorporate its' file type
as well. This is later used to convert the file back into the base64
string and passed to JavaScript to preview the image while it is being
converted.

Save the converted file in chosen location

Body no longer has display flex. Header has an explicit height and main
uses that to determine its' own height. The functionality is still the
same but now main won't flex grow out of the screen.

Added class to fileInputBox to reference it in css. Now css doesn't use
any ids. This reduces the chance of specificity issues.