Bug fixes and minor code changes
History is now read and displayed in reverse order, with the most recent
at the top

Moved code to make history directory to when the program is run. This
stops an error being thrown when the user views history before
converting.

Fixed bug where file input was still enabled when saving files