Improve input constraint checking on client side
Renamed setting css class to settingDiv

settings.js now stores read settings in variable instead of passing it
multiple times to several different functions

Input constraints are also enforced manually before updating the
settings. The HTML input constraints don't work if the user types in a
value.
 - The dictionary path is already resolved from backtracking through
   parent elements and using their ids as keys. The path is used in
   combination with the full SETTINGS variable to find the section.
 - The immediate dictionary is passed by reference so the immediate
   value can still be modified, when checking for constraints, after the
   full, back-tracked dictionary has been constructed.

Changed string inputs to dropdown menus and added constraint of
supported fonts for FPDF

Constraints are also first looked for in the same section then in the
--constraints section of the full SETTINGS variable

-- had been changed to __ in the settings dictionary in order to be more
compliant with JavaScript's parsing of the object