Fix bugs, add fullSettings variable
Fixed bug where document preferences weren't being used when creating
the document

Fixed bug where the text was in unicode instead of latin-1 encoding for
the fpdf library

Fixed bug where settings weren't being read from the file, it was due to
a typo

Fixed bug where an empty alert would display even if there was no
errorMsg about the uploaded files

Added fullSettings variable that contains userSettings and
defaultSettings

Added failsafe to remove toConvert directory before adding to it. This
should never be needed as the directory is always removed after it has
been used but if the program crashes it may remain.