Store all documents in sqlite database
Upon run the database file is created if it doesn't exist, its tables
are created if they don't exist, and document records without a creation
time are deleted as this means that they never finished conversion

Database structure:
 - Table document
   - id
   - save_location
   - creation_time
 - image
   - id
   - document_id
   - position
   - data
   - filetype

Inputted documents are stored in the document table with a Null
creation_time which is filled in after conversion. This also helps
identify which documents haven't finished conversion so they be deleted on
run if they exist or be accessed for conversion.

Inputted images are stored in the image table with the document_id
foreign key

Reading history reads all documents sorted by descending creation time
and then reads the images with the same document_id

Clearing history deletes all records from both tables

Removed multiple functions to escape and unescape filenames and to
create data URIs

Removed shutil import as it is no longer needed as there is no toConvert
directory to delete

Added function to create data URI from passed filetype and data

Modified save function to save the file directly to its final location

Modified ai.py to use base64 string instead of file from disk. Numpy
converts it to an array and OpenCV converts that to an image object.

Modified uploadFiles.js to only pass the data URI for each image not the
file type as well

Modified .gitignore to ignore database file and removed history and
toConvert directory