Fix bugs, add settings verification, optimisations
Removed 'All files' option from save location picker, I think this may
cause problems on Windows but it works fine on MacOS

__main__.py now adds default file extension to save location if user
picks a location without specifying one

In ai.py moved all formatting code into the appropriate function

Removed a double space in converting.css

Changed some spaces to tabs in index.html

In settings.js I changed the settings elements so now the key string is
stored in the data-key attribute rather than the id. This is because
there may be multiple settings in different sub dictionaries with the
same name but there can only be one of an id on the page.

Inputs are still given ids for the labels to identify them. This is
created by a function that just increments and returns an id each time
it is called.

Added verification for string inputs given via select tag to settings.js

Fixed bug where files will be input in the wrong order. This was due to
them being processed asynchronously to each other. Now instead of
starting the input of each file via iteration, each file recursively
inputs the next file until all files are inputted. This maintains their
order.